requirements.txt
Fabric
Flask-SQLAlchemy
Flask-WTF
WTForms
coverage
python-chess
futures
flask_cors
# ecdsa
# paramiko
# pycrypto
